backend/requirements.txt
Django>=5.1,<5.3
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.4
psycopg[binary]>=3.2
python-decouple>=3.8
channels>=4.1
channels-redis>=4.2
celery>=5.4
redis>=5.0
Pillow>=10.4
cloudinary>=1.41
django-cloudinary-storage>=0.3.0

